+++ settings.js
@@ -43,6 +43,12 @@ export const registerSettings = function () {
 		'gm': game.i18n.localize("MonksTokenBar.GMOnly"),
 		'players': game.i18n.localize("MonksTokenBar.PlayersOnly"),
 		'everyone': game.i18n.localize("MonksTokenBar.Everyone"),
+	}
+
+	let permissions = {
+		"LIMITED": "Limited",
+		"OBSERVER": "Observer",
+		"OWNER": "Owner"
     }
 
 	let lootsheetoptions = MonksTokenBar.getLootSheetOptions();
@@ -79,6 +85,15 @@ export const registerSettings = function () {
 		config: true,
 		default: false,
 		type: Boolean,
+	});
+	game.settings.register(modulename, "minimum-ownership", {
+		name: game.i18n.localize("MonksTokenBar.minimum-ownership.name"),
+		hint: game.i18n.localize("MonksTokenBar.minimum-ownership.hint"),
+		scope: "world",
+		config: true,
+		default: "OWNER",
+		type: String,
+		choices: permissions,
 	});
 	game.settings.register(modulename, "disable-tokenbar", {
 		name: game.i18n.localize("MonksTokenBar.disable-tokenbar.name"),
@@ -262,6 +277,15 @@ export const registerSettings = function () {
 		localize: true
 	});
 
+	game.settings.register(modulename, "show-lootable-menu", {
+		name: game.i18n.localize("MonksTokenBar.show-lootable-menu.name"),
+		hint: game.i18n.localize("MonksTokenBar.show-lootable-menu.hint"),
+		scope: "world",
+		config: true,
+		default: true,
+		type: Boolean,
+	});
+
 	game.settings.register(modulename, "gold-formula", {
 		name: game.i18n.localize("MonksTokenBar.gold-formula.name"),
 		hint: game.i18n.localize("MonksTokenBar.gold-formula.hint"),
@@ -269,6 +293,14 @@ export const registerSettings = function () {
 		config: true,
 		default: "Math.round(0.6 * 10 * (10 ** (0.15 * ({{ actor.system.details.cr}} ?? 0))))",
 		type: String,
+	});
+	game.settings.register(modulename, "auto-gold-cr", {
+		name: game.i18n.localize("MonksTokenBar.auto-gold-cr.name"),
+		hint: game.i18n.localize("MonksTokenBar.auto-gold-cr.hint"),
+		scope: "world",
+		config: true,
+		default: false,
+		type: Boolean,
 	});
 
 	game.settings.register(modulename, "loot-sheet", {
@@ -288,15 +320,6 @@ export const registerSettings = function () {
 		default: "",
 		type: String,
 	});
-	game.settings.register(modulename, "loot-folder", {
-		name: game.i18n.localize("MonksTokenBar.loot-folder.name"),
-		hint: game.i18n.localize("MonksTokenBar.loot-folder.hint"),
-		scope: "world",
-		config: false,
-		default: "",
-		choices: lootfolder,
-		type: String,
-	});
 	game.settings.register(modulename, "create-canvas-object", {
 		name: game.i18n.localize("MonksTokenBar.create-canvas-object.name"),
 		hint: game.i18n.localize("MonksTokenBar.create-canvas-object.hint"),
@@ -313,14 +336,6 @@ export const registerSettings = function () {
 		default: "none",
 		choices: openLootOptions,
 		type: String,
-	});
-	game.settings.register(modulename, "show-lootable-menu", {
-		name: game.i18n.localize("MonksTokenBar.show-lootable-menu.name"),
-		hint: game.i18n.localize("MonksTokenBar.show-lootable-menu.hint"),
-		scope: "world",
-		config: true,
-		default: true,
-		type: Boolean,
 	});
 
 	//------------------------------------Request Roll settings--------------------------------------------
@@ -353,6 +368,14 @@ export const registerSettings = function () {
 		hint: game.i18n.localize("MonksTokenBar.gm-sound.hint"),
 		scope: "world",
 		config: !game.modules.get("dice-so-nice")?.active,
+		default: true,
+		type: Boolean,
+	});
+	game.settings.register(modulename, "add-advantage-buttons", {
+		name: game.i18n.localize("MonksTokenBar.add-advantage-buttons.name"),
+		hint: game.i18n.localize("MonksTokenBar.add-advantage-buttons.hint"),
+		scope: "world",
+		config: true,
 		default: true,
 		type: Boolean,
 	});